+++ src/Akka.Hosting.Tests/Logging/SemanticLoggingSpecs.cs
@@ -377,7 +377,7 @@ public class SemanticTestLogger : ILogger
 
     public bool IsEnabled(LogLevel logLevel) => true;
 
-    public IDisposable BeginScope<TState>(TState state) => EmptyDisposable.Instance;
+    public IDisposable? BeginScope<TState>(TState state) where TState : notnull => EmptyDisposable.Instance;
 }
 
 public class LogEntry